requirements.txt
pandas==2.2.0
pyarrow==15.0.0
boto3==1.34.0
sqlalchemy==2.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
streamlit==1.56.0
plotly==5.24.0
email-validator==2.2.0